Adicionar pydantic
