+++ pyproject.toml
@@ -13,6 +13,7 @@ uvicorn = "^0.24.0.post1"
 taskipy = "^1.12.2"
 pytest = "^7.4.3"
 httpx = "^0.25.2"
+pydantic = "^2.5.2"
 
 
 [build-system]
@@ -21,6 +22,6 @@ build-backend = "poetry.core.masonry.api"
 
 [tool.taskipy.tasks]
 
-run = "uvicorn main:app --reload"
+run = "uvicorn app.main:app --reload"
 test = "pytest -v"
 kill = "kill -9 $(lsof -t -i:8000)"
